LayoutTests/fast/events/dispatch-event-being-dispatched-expected.txt

Tests that dispatchEvent raises INVALID_STATE_ERR if the event being dispatched is already being dispatched.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.dispatchEvent(event) threw exception Error: InvalidStateError: DOM Exception 11.
PASS redispatchCustom.wasInvoked is true
PASS window.dispatchEvent(event) threw exception Error: InvalidStateError: DOM Exception 11.
PASS checkCustom.wasInvoked is true
PASS document.dispatchEvent(event) threw exception Error: InvalidStateError: DOM Exception 11.
PASS successfullyParsed is true

TEST COMPLETE
